requirements.txt
google-generativeai>=0.3.0
redis>=4.5.0
prometheus-client>=0.17.0
flask>=2.3.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
psutil>=5.9.0
gputil>=1.4.0
cryptography>=39.0.0
dnspython>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3